src/app/modules/valorisation-organique-economie-circulaire/ai-model/nutriflow-onnx-api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
onnxruntime>=1.16.0
numpy>=1.24.0
Pillow>=10.0.0
